cleanup and shutdown management
