+++ nym-authenticator-client/Cargo.toml
@@ -25,8 +25,8 @@ nym-authenticator-requests = { path = "../common/authenticator-requests" }
 nym-bandwidth-controller = { path = "../common/bandwidth-controller" }
 nym-credentials-interface = { path = "../common/credentials-interface" }
 nym-crypto = { path = "../common/crypto" }
-nym-node-requests = { path = "../nym-node/nym-node-requests" }
 nym-pemstore = { path = "../common/pemstore" }
+nym-registration-common = { path = "../common/registration" }
 nym-sdk = { path = "../sdk/rust/nym-sdk" }
 nym-service-provider-requests-common = { path = "../common/service-provider-requests-common" }
 nym-validator-client = { path = "../common/client-libs/validator-client" }
