+++ nym-registration-client/Cargo.toml
@@ -19,11 +19,10 @@ tracing.workspace = true
 url.workspace = true
 
 nym-authenticator-client = { path = "../nym-authenticator-client" }
-nym-authenticator-requests = { path = "../common/authenticator-requests" }
 nym-bandwidth-controller = { path = "../common/bandwidth-controller" }
 nym-credential-storage = { path = "../common/credential-storage" }
 nym-credentials-interface = { path = "../common/credentials-interface" }
 nym-ip-packet-client = { path = "../nym-ip-packet-client" }
-nym-ip-packet-requests = { path = "../common/ip-packet-requests" }
+nym-registration-common = { path = "../common/registration" }
 nym-sdk = { path = "../sdk/rust/nym-sdk" }
 nym-validator-client = { path = "../common/client-libs/validator-client" }
