+++ nym-registration-client/src/builder/mod.rs
@@ -5,7 +5,7 @@ use nym_bandwidth_controller::{BandwidthController, BandwidthTicketProvider};
 use nym_credential_storage::ephemeral_storage::EphemeralCredentialStorage;
 use nym_sdk::{
     mixnet::{MixnetClient, MixnetClientBuilder},
-    NymNetworkDetails, ShutdownManager,
+    NymNetworkDetails,
 };
 use nym_validator_client::{
     nyxd::{Config as NyxdClientConfig, NyxdClient},
@@ -38,7 +38,6 @@ impl RegistrationClientBuilder {
             data_path: self.config.data_path.clone(),
         };
         let cancel_token = self.config.cancel_token.clone();
-        let mixnet_shutdown_manager = ShutdownManager::build_new_default()?;
 
         let nyxd_client = get_nyxd_client(&self.config.network_env)?;
 
@@ -46,8 +45,7 @@ impl RegistrationClientBuilder {
             MixnetClient,
             Box<dyn BandwidthTicketProvider>,
         ) = if let Some((mixnet_client_storage, credential_storage)) = storage {
-            let builder = MixnetClientBuilder::new_with_storage(mixnet_client_storage)
-                .custom_shutdown(mixnet_shutdown_manager.shutdown_tracker().clone());
+            let builder = MixnetClientBuilder::new_with_storage(mixnet_client_storage);
             let mixnet_client = tokio::time::timeout(
                 MIXNET_CLIENT_STARTUP_TIMEOUT,
                 self.config.build_and_connect_mixnet_client(builder),
@@ -57,8 +55,7 @@ impl RegistrationClientBuilder {
                 Box::new(BandwidthController::new(credential_storage, nyxd_client));
             (mixnet_client, bandwidth_controller)
         } else {
-            let builder = MixnetClientBuilder::new_ephemeral()
-                .custom_shutdown(mixnet_shutdown_manager.shutdown_tracker().clone());
+            let builder = MixnetClientBuilder::new_ephemeral();
             let mixnet_client = tokio::time::timeout(
                 MIXNET_CLIENT_STARTUP_TIMEOUT,
                 self.config.build_and_connect_mixnet_client(builder),
@@ -76,7 +73,6 @@ impl RegistrationClientBuilder {
             mixnet_client,
             config,
             cancel_token,
-            mixnet_shutdown_manager,
             mixnet_client_address,
             bandwidth_controller,
         })
