+++ nym-registration-client/src/lib.rs
@@ -1,42 +1,32 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
-use std::net::IpAddr;
 
-use nym_authenticator_client::{
-    AuthClientMixnetListener, AuthClientMixnetListenerHandle, AuthenticatorClient,
-};
-use nym_bandwidth_controller::BandwidthTicketProvider;
-use nym_credentials_interface::TicketType;
-use nym_sdk::mixnet::{MixnetClient, Recipient};
-use nym_sdk::ShutdownManager;
 use tokio_util::sync::CancellationToken;
 
+use nym_authenticator_client::{AuthClientMixnetListener, AuthenticatorClient};
+use nym_bandwidth_controller::BandwidthTicketProvider;
+use nym_credentials_interface::TicketType;
 use nym_ip_packet_client::IprClientConnect;
-use nym_ip_packet_requests::IpPair;
+use nym_registration_common::AssignedAddresses;
+use nym_sdk::mixnet::{MixnetClient, Recipient};
 
 use crate::config::RegistrationClientConfig;
 
 mod builder;
 mod config;
 mod error;
+mod types;
 
 pub use builder::config::{BuilderConfig as RegistrationClientBuilderConfig, MixnetClientConfig};
 pub use builder::RegistrationClientBuilder;
-pub use config::NymNode as RegistrationClientNymNode;
 pub use error::RegistrationClientError;
-
-pub use nym_authenticator_client::{
-    GatewayData, DEFAULT_PRIVATE_ENTRY_WIREGUARD_KEY_FILENAME,
-    DEFAULT_PRIVATE_EXIT_WIREGUARD_KEY_FILENAME, DEFAULT_PUBLIC_ENTRY_WIREGUARD_KEY_FILENAME,
-    DEFAULT_PUBLIC_EXIT_WIREGUARD_KEY_FILENAME,
-};
+pub use types::{MixnetRegistrationResult, RegistrationResult, WireguardRegistrationResult};
 
 pub struct RegistrationClient {
     mixnet_client: MixnetClient,
     config: RegistrationClientConfig,
     mixnet_client_address: Recipient,
     bandwidth_controller: Box<dyn BandwidthTicketProvider>,
-    mixnet_shutdown_manager: ShutdownManager,
     cancel_token: CancellationToken,
 }
 
@@ -61,7 +51,6 @@ impl RegistrationClient {
         Ok(RegistrationResult::Mixnet(Box::new(
             MixnetRegistrationResult {
                 mixnet_client: ipr_client.into_mixnet_client(),
-                mixnet_shutdown_manager: self.mixnet_shutdown_manager,
                 assigned_addresses: AssignedAddresses {
                     interface_addresses,
                     exit_mix_address: ipr_address,
@@ -93,12 +82,8 @@ impl RegistrationClient {
 
         // Start the auth client mixnet listener, which will listen for incoming messages from the
         // mixnet and rebroadcast them to the auth clients.
-        let mixnet_listener = AuthClientMixnetListener::new(
-            self.mixnet_client,
-            self.cancel_token.clone(),
-            self.mixnet_shutdown_manager,
-        )
-        .start();
+        let mixnet_listener =
+            AuthClientMixnetListener::new(self.mixnet_client, self.cancel_token.clone()).start();
 
         let mut entry_auth_client = AuthenticatorClient::new_entry(
             &self.config.data_path,
@@ -154,33 +139,4 @@ impl RegistrationClient {
             .await
             .ok_or(RegistrationClientError::Cancelled)?
     }
-}
-
-pub enum RegistrationResult {
-    Mixnet(Box<MixnetRegistrationResult>),
-    Wireguard(Box<WireguardRegistrationResult>),
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct AssignedAddresses {
-    pub entry_mixnet_gateway_ip: IpAddr,
-    pub exit_mixnet_gateway_ip: IpAddr,
-    pub mixnet_client_address: Recipient,
-    pub exit_mix_address: Recipient,
-    pub interface_addresses: IpPair,
-}
-
-pub struct MixnetRegistrationResult {
-    pub assigned_addresses: AssignedAddresses,
-    pub mixnet_client: nym_sdk::mixnet::MixnetClient,
-    pub mixnet_shutdown_manager: nym_sdk::ShutdownManager,
-}
-
-pub struct WireguardRegistrationResult {
-    pub entry_gateway_client: AuthenticatorClient,
-    pub exit_gateway_client: AuthenticatorClient,
-    pub entry_gateway_data: GatewayData,
-    pub exit_gateway_data: GatewayData,
-    pub authenticator_listener_handle: AuthClientMixnetListenerHandle,
-    pub bw_controller: Box<dyn BandwidthTicketProvider>,
 }
